requirements.txt
# REQUIRED
django >=3.2.13,<3.3
django-reversion
psycopg2-binary
uwsgi
pyshp

# 3rd Party Form Fields
django-phone-field

# TESTING
coverage

#OPTIONAL, BUT LOVELY FOR DEBUGGING
ipdb
ipython